# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/plants/{id}": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get plant by ID",
                "operationId": "59659eecce74ef80d892d5567298021e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Plant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plant details"
                    },
                    "404": {
                        "description": "Plant not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Plants"
                ],
                "summary": "Edit a plant",
                "operationId": "f6d7c3d40cedd8da8f5f2a67dbd0a03b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Plant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated plant data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "location_id": {
                                        "type": "integer"
                                    },
                                    "trefle_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "desc": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plant updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Plant not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Plants"
                ],
                "summary": "Delete a plant",
                "operationId": "98060208c3a7155b244377f43b937076",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Plant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plant deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Plant not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/plants/{idLocation}/plants": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get plants by location ID",
                "operationId": "5c1a54f9913c4e10dea7b318fb01ba16",
                "parameters": [
                    {
                        "name": "idLocation",
                        "in": "path",
                        "description": "Location ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of plants at the location"
                    },
                    "404": {
                        "description": "No plants found at the location"
                    }
                }
            }
        },
        "/plants": {
            "post": {
                "tags": [
                    "Plants"
                ],
                "summary": "Create a new plant",
                "operationId": "d1e698cfef1502cd76c06d34296f9ef9",
                "requestBody": {
                    "description": "Plant data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "location_id": {
                                        "type": "integer"
                                    },
                                    "trefle_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "desc": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plant created successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Location not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}